Enable Shop Creation Selector By Default for All Users
Summary:
There have been no negative reports about the shops. It's time to enable it for everyone. Previously it was hidden in advanced settings (and set to off by default), now it's default on.
I updated the main text in the plugin page to reflect this.
This will be changing soon anyway so I don't think we should trouble Content/Design with this since they are revamping this completely in the next few weeks.

Reviewed By: jieminz22

Differential Revision: D7674848

fbshipit-source-id: b76ae24