Enable support for the 'reset' message.
Summary:
Before, we didn't do anything on reset (when user hits delete inside FAE, a reset message is sent), because the design was flexibile enough to accomodate that.

Now we will reset the merchant settings id.
If we don't do this, the new design will not work correctly, it will display a button to `Create Ad` but that button won't work because the EMS is gone.

Reviewed By: mengyingdu

Differential Revision: D8241691

fbshipit-source-id: dc5fa06