Detect Problems and Surface Custom Link to Debug Mode
Summary:
I added a setTimeout which will fire after 10 seconds and check if there is a problem with the loaded params, if there is a problem, the following screen will be displayed.
{F132250423}

I also added a loading screen. The purpose of this is the user knows to wait. Otherwise the screen is just blank (as it is in all our GitHub issues on this).
{F132250422}

Reviewed By: mengyingdu

Differential Revision: D8796137

fbshipit-source-id: f13cbef3