Fix AddToCart Matching Error When adding products to cart that arent in FB Product Catalog
Summary:
When adding simple products (variants of a configurable product) into the cart in some stores the ID that the code is getting is the id of the simple product, which is not imported individually into FB if it is marked as invisible.
This diff gets the ID of the configurable product and adds it to the cart instead of adding the simple product to the cart which causes a matching error.

Reviewed By: mengyingdu

Differential Revision: D9071369

fbshipit-source-id: ca199257