Debug Mode - Part 3 : The Final Form
Summary:
{F134793167}
{F134793168}

The Debug Mode now consists of a button which generates log files.
Then the UI provides instructions on what to do, which is Go to GitHub and post the log files.

The button tries to generate a sample feed (I'm pretty sure most of the problems are there, but if they aren't we can do a part 4 and test other things too).
I added debugging code to the sample feed generation conditional on the debugging param.
I tried to minimize changes to production and so I used a mock object `FBDebugProduct` and used
the php magic `__call` method to simulate a Magento product while logging every access.

Every access to this object is sent to a log file, so we will know exactly where the problem is (the last un-successfull log entry).
I also figured out a way to log errors that are impossible to log (normally), this can be done by hooking into the shutdown method via `register_shutdown_function` then logging the last error.

Reviewed By: mengyingdu

Differential Revision: D9174475

fbshipit-source-id: b3ea8bb1