Fix Broken Inheritance due to Method Signature Change - PHP 5.3
Summary:
I found this by testing on a PHP 5.3 server for aws.
Essentially, I changed the method signature of a key method but forgot about the other files that inherited it.

The new way to avoid changing the signature is to use `$this->stock` essentially the problem I have here is I need to spoof the stock item.

Reviewed By: mengyingdu

Differential Revision: D9302886

fbshipit-source-id: 3e873581