Fixes Incorrectly In-Stock Configurable Products
Summary:
This fixes the case where all children of a configurable product are out of stock,
but the parent is still marked in stock. This will probably result in a performance hit but should fix
issue #6 on GitHub.

The fix is done by calling the `isAvailable` method on the product, which is what the frontend uses to determine
whether to display in stock or out of stock, so it is what FB should be using as well.

Reviewed By: mengyingdu

Differential Revision: D13326421

fbshipit-source-id: 50cd752f