Adding contents field to purchase pixel events
Summary:
The contents field is sent in Conversions API events when a user makes a purchase, but it was missing on pixel events: https://developers.facebook.com/docs/facebook-pixel/reference#standard-events.
A new method is added in the associated block, and it's called in the corresponding template. The regular expression in the functional test is changed as well.

Reviewed By: vinothsa

Differential Revision: D22638497

fbshipit-source-id: db300e39