Adding missing fields to initiate checkout pixel event
Summary:
The contents and num_items field are sent in Conversions API events when a user initiates checkout, but they were missing on pixel events: https://developers.facebook.com/docs/facebook-pixel/reference#standard-events.
A pair of methods were added in the associated block, and are called in the corresponding template. The regular expression in the functional test is changed too.

Reviewed By: vinothsa

Differential Revision: D22638201

fbshipit-source-id: e2172eb0